pyproject.toml
[tool.poetry]
authors = ["BPR Authors"]
description = ""
name = "bpr"
version = "0.0.1"

[tool.poetry.dependencies]
comet-ml = "^3.1.14"
faiss-cpu = "^1.6.5"
joblib = "^1.0.0"
lmdb = "^1.0.0"
python = "^3.7"
pytorch-lightning = "0.9.0"
regex = "^2020.11.13"
torch = "1.6.0"
tqdm = "4.46.1"
transformers = "2.11.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
mypy = "^0.782"
mypy-extensions = "0.4.3"

[tool.black]
line-length = 120

[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]
